requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
numpy
pymupdf
pytest
httpx
python-multipart
google-generativeai
sentence-transformers
transformers
accelerate
bitsandbytes
streamlit
plotly
faiss-cpu>=1.7.0
psutil
torch
